3  more questions 3

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long int a;
 	long long int b;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(c));
-	printf("Size of an int: %lu byte(s)\n", sizeof(i));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(a));
-	printf("Size of a long long: %lu byte(s)\n", sizeof(b));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n",(unsigned long)sizeof(c));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(i));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of a long long: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
